refactor: separa as classes em modulos
